### 2023-02-02 #80 @skrawcz
I'm still working on the JOSS piece, but should have that done this week.

> The package is deposited in a long-term repository with the DOI:
Though I'm not entirely sure how to fulfill this requirement -- I haven't search for the answer yet, but if you have a link that would be appreciated.
