### 2023-02-02 #80 @NickleDave
Hi @skrawcz and thank you for taking the time to make the full submission.

I need to ask about one aspect of Hamilton we did not discuss before.

I see the developers recently added opt out telemetry. Two questions:
- Are users given any notification that telemetry is being used if they do **not** opt out? E.g., with a print statement, through `logging`. or `warnings.warn`? This was not clear to me from the README.
- I am really sympathetic to the devs wanting simple usage stats, and the language in the README makes it sound like you want to limit what information is collected to just this anonymous data. But orgs similar to ours usually require that any telemetry be opt in (e.g. [Linux Foundation](https://www.linuxfoundation.org/legal/telemetry-data-policy)). How open would you all be to making the telemetry opt in instead?
