### 2023-02-03 #80 @NickleDave
> No, but I just added it - [stitchfix/hamilton#302](https://github.com/stitchfix/hamilton/pull/302).

Thank you 🙏 I think this goes a long way towards alleviating our concerns, and is similar to what [Homebrew does](https://github.com/Homebrew/brew/blob/master/docs/Analytics.md#anonymous-aggregate-user-behaviour-analytics) (as you may already know).

> By having it opt-out (we've only had it on for a month) it's already been valuable ...  and ... is helping inform us what parts of the library are being used

I hear you, believe me.

As you can probably tell, this is a case where we did not yet have policy in place. We need to be sure we're doing what would be acceptable to the rest of the scientific Python community here, I hope you can understand. Please let me update you tomorrow afternoon at the latest.
