### 2023-02-05 25d9e87
added HamiltonTransformer class w/ basic checks
