+++ examples/scikit-learn/run.py
@@ -1,0 +1,145 @@
+import importlib
+from types import ModuleType
+from typing import Any, Dict, List
+
+import numpy as np
+import pandas as pd
+from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.pipeline import Pipeline
+from sklearn.preprocessing import StandardScaler
+from sklearn.utils.validation import check_array, check_is_fitted
+
+from hamilton import base, driver, log_setup
+
+
+class HamiltonTransformer(BaseEstimator, TransformerMixin):
+    """Scikit-learn compatible Transformer implementing Hamilton behavior"""
+
+    def __init__(
+        self,
+        config: dict = None,
+        modules: List[ModuleType] = None,
+        adapter: base.HamiltonGraphAdapter = None,
+        final_vars: List[str] = None,
+        overrides: Dict[str, Any] = None,
+    ):
+        self.config = {} if config is None else config
+        self.modules = [] if modules is None else modules
+        self.adapter = adapter
+        self.final_vars = [] if final_vars is None else final_vars
+        self.overrides = overrides
+
+    def get_params(self):
+        return {
+            "config": self.config,
+            "modules": self.modules,
+            "adapter": self.adapter,
+            "final_vars": self.final_vars,
+            "overrides": self.overrides,
+        }
+
+    def set_params(self, **parameters):
+        for parameter, value in parameters.items():
+            setattr(self, parameter, value)
+        return self
+
+    def get_features_names_out(self):
+        if self.feature_names_out_:
+            return self.feature_names_out_
+
+    def _get_tags(self):
+        return {"requires_fit": False, "requires_y": False}
+
+    def fit(self, X, y=None):
+        check_array(X, accept_sparse=True)
+        self.driver_ = driver.Driver(self.config, *self.modules, adapter=self.adapter)
+        self.n_features_in_ = X.shape[1]
+
+        return self
+
+    def transform(self, X, y=None, **kwargs):
+        check_is_fitted(self, "n_features_in_")
+
+        if isinstance(X, pd.DataFrame):
+            check_array(X, accept_sparse=True)
+            if X.shape[1] != self.n_features_in_:
+                raise ValueError("Shape of input is different from what was seen in `fit`")
+
+            X = X.to_dict(orient="series")
+
+        X_t = self.driver_.execute(final_vars=self.final_vars, overrides=self.overrides, inputs=X)
+
+        self.n_features_out_ = len(self.final_vars)
+        self.feature_names_out_ = X_t.columns.to_list()
+        return X_t
+
+    def fit_transform(self, X, y=None, **fit_params):
+        return self.fit(X, **fit_params).transform(X)
+
+
+if __name__ == "__main__":
+    log_setup.setup_logging()
+
+    module_names = ["my_functions_a", "my_functions_b"]
+    modules = [importlib.import_module(m) for m in module_names]
+
+    initial_df = pd.DataFrame(
+        {"signups": [1, 10, 50, 100, 200, 400], "spend": [10, 10, 20, 40, 40, 50]}
+    )
+
+    output_columns = [
+        "spend",
+        "signups",
+        "avg_3wk_spend",
+        "spend_per_signup",
+        "spend_zero_mean_unit_variance",
+    ]
+
+    # Check 1: output of `vanilla driver` == `custom transformer`
+    dr = driver.Driver({}, *modules)
+    hamilton_df = dr.execute(final_vars=output_columns, inputs=initial_df.to_dict(orient="series"))
+
[…]
