### 2023-02-09 4c2a4a4
Update run.py
Make tag coherent with the design decision to have driver instantiated in .fit()
