+++ examples/scikit-learn/run.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+
 import importlib
 import logging
 from types import ModuleType
