+++ docs/best-practices/using-within-your-etl-system.rst
@@ -25,7 +25,7 @@ Compatibility Matrix
    * - `dbt <https://getdbt.com/>`_
      - ❔  (dbt did not run python, but now it does so it should)
    * - `kubernetes <https://kubernetes.io/>`_
-     - ✅ but you need to setup kubernetes to run an image that can run python code - e.g. see `https://medium.com/avmconsulting-blog/running-a-python-application-on-kubernetes-aws-56609e7cd88c <https://medium.com/avmconsulting-blog/running-a-python-application-on-kubernetes-aws-56609e7cd88c>`_
+     - ✅ but you need to setup kubernetes to run an image that can run python code - e.g. see `Running a python application on kubernetes <https://medium.com/avmconsulting-blog/running-a-python-application-on-kubernetes-aws-56609e7cd88c/>`_
    * - `docker <https://www.docker.com/>`_
      - ✅ but you need to setup a docker image that can execute python code.
    * - ... in general if it runs python 3.6+ ...
