### 2023-02-14 cfb9963
remove .vscode files and update links
