+++ docs/contributing.rst
@@ -4,7 +4,7 @@ Contributing
 
 For contributing guidelines see github.
 
-We open contributions big and small. See our `contributing guidelines <https://github.com/stitchfix/hamilton/blob/main/CONTRIBUTING.md/>`_.
+We open contributions big and small. See our `contributing guidelines <https://github.com/stitchfix/hamilton/blob/main/CONTRIBUTING.md>`_.
 
-We also operate under a `Code of Conduct <https://github.com/stitchfix/hamilton/blob/main/CODE\_OF\_CONDUCT.md/>`_, and
+We also operate under a `Code of Conduct <https://github.com/stitchfix/hamilton/blob/main/CODE\_OF\_CONDUCT.md>`_, and
 expect contributors to do the same.
