### 2023-02-15 #80 @lwasser
open issue is here - https://github.com/pyOpenSci/software-peer-review/issues/183
