### 2023-02-16 #80 @skrawcz
Awesome, sounds good @lwasser. I might be a bit slow, but will try to pen some thoughts. Will send you an email too. Cheers.
