### 2023-02-17 484678f
Adds documentation for decorators
We get at the mental model for decorators/how layering works. Its
complicated, so we don't go into too much detail (the code is there to
browse), but should give users enough to be effective at using
layering/`target_`
