+++ decorators.md
@@ -285,8 +285,10 @@ dr = driver.Driver(config, module1, ...)
 
 ### @tag
 
-Allows you to attach metadata to an output(s), i.e. all nodes generated by a function and its decorators (note, this only applies to "final" nodes --
-not any intermediate nodes that are generated...).
+Allows you to attach metadata to an output(s), i.e. all nodes generated by a function and its decorators (note, this, by default, only applies to "final" nodes --
+not any intermediate nodes that are generated...). See "A mental model for decorators" below for how/why this works.
+
+```python
 A common use of this is to enable marking nodes as part of some data product, or for GDPR/privacy purposes.
 
 For instance:
@@ -302,7 +304,11 @@ def intermediate_column() -> pd.Series:
 def final_column(intermediate_column: pd.Series) -> pd.Series:
     pass
 ```
-### @tag_outputs
+
+`@tag` also allows you to specify a target with the `target_` parameter. See "A mental model for decorators" below for more details.
+
+
+### @tag_outputs (deprecated)
 
 `tag_outputs` enables you to attach metadata to a function that outputs multiple nodes,
 and give different tag values to different outputs:
@@ -387,7 +393,9 @@ The check_output validator takes in arguments that each correspond to one of the
 These arguments tell it to add the default validator to the list. The above thus creates
 two validators, one that checks the datatype of the series, and one that checks whether the data is in a certain range.
 
+
 Note that you can also specify custom decorators using the `@check_output_custom` decorator.
+`check_output` and `check_output_custom` accept the `_target` parameter, which specifies *which* output to check.
 
 See [data_quality](data_quality.md) for more information on available validators and how to build custom ones.
 
@@ -558,3 +566,81 @@ Models (optionally) accept a `output_column` parameter -- this is specifically i
 from the output column that it should represent. E.G. if you use the model result as an intermediate object, and manipulate
 it all later. At Stitch Fix this is necessary because various dependent columns that a model queries
 (e.g. `MULTIPLIER_...` and `OFFSET_...`) are derived from the model's name.
+
+## A mental model for decorators
+
+Decorators fall under a few broad categories:
+
+1. Decorators that decide whether a function should resolves into a node (E.G. config.when())
+2. Decorators that *turn* a function into a set of nodes (E.G. `@does`)
+3. Decorators that modify a set of nodes, turning it into another set of nodes (E.G. `@parameterize`, `@extract_columns`, and `@check_output`)
+
+These are broken up into subclasses that each have various nuances, one of which is whether or not they allow layering.
+To layer effectively, we need to know *which* nodes a decorator will modify from the result of another decorators. For the ones
+that allow this type of layers, they take in a `target_` parameter (with `_` due to the fact that they often have **kwargs) that tells *which*
+nodes they should modify.
+
+There are 4 options for `target_`:
+
+1. None -- this is the default behavior. This only applies to nodes generated by the function that are not
+depended on by any other nodes generated by the function.
+2. `...` (the literal python Ellipsis). This means that *all* nodes generated by the function and its decorators
+   will have the transformation applied to them.
+3. A string. This means that only the node with the specified name will have the transformation applied.
+4. A list of strings. This means that all nodes with the specified name will have the transformation applied.
+
+This is very powerful, and we will likely be applying it to more decorators. Currently `tag` and `check_output` are the only
+decorators that accept this parameter.
+
+To dive into this, let's take a look at the `tag` decorator with the default `target_`:
+
+This will only decorate `col1`, `col2`, and `col3`:
+
[…]
