### 2023-02-19 985819a
Adds code to make concatenating dataframes in the result builder default
Why this change? because if you are doing indicator variable stuff,
maybe you don't want to be explicit about them, instead you want
to concatentate the columns created without knowing about them. Seems
like a reasonable pattern on first glance.

Otherwise this isn't pretty code. Still thinking it through, but essentially:

1. want to handle column concat of dataframes and series.
2. problem is that pd.DataFrame({ dict }) behavior essentially does
an outer join on indexes if present, and handles scalars.
3. also we need t
[…]
