### 2023-02-20 f3eee93
Simplifies duplicate column check in dataframe builder
No need for a separate data structure. Can just use one.
