+++ hamilton/base.py
@@ -189,33 +189,25 @@ class PandasDataFrameResult(ResultMixin):
         :param outputs: The outputs to build the dataframe from.
         :return: A dataframe with the outputs.
         """
+
+        def get_output_name(output_name: str, column_name: str) -> str:
+            """Add function prefix to columns.
+            Note this means that they stop being valid python identifiers due to the `.` in the string.
+            """
+            return f"{output_name}.{column_name}"
+
         flattened_outputs = {}
         for name, output in outputs.items():
             if isinstance(output, pd.DataFrame):
-                df_columns = list(output.columns)
-                column_intersection = [
-                    column for column in df_columns if column in flattened_outputs
-                ]
-                if column_intersection:
-                    raise ValueError(
-                        f"Dataframe {name} contains columns {column_intersection} that already exist in the output. "
-                        f"Please rename the columns in {name} to avoid this error."
-                    )
                 if logger.isEnabledFor(logging.DEBUG):
                     logger.debug(
-                        f"Unpacking dataframe {name} into dict of series with columns {df_columns}."
+                        f"Unpacking dataframe {name} into dict of series with columns {list(output.columns)}."
                     )
-                name_prefix = f'{name.split("_")[0]}_'
-                if any([column for column in df_columns if not column.startswith(name_prefix)]):
-                    # adding this here to try to force some consistency in the naming of columns and mapping it to the
-                    # originating function.
-                    logger.warning(
-                        f"Friendly warning: you're unpacking the dataframe from function [{name}()] into a dict of "
-                        f"series with columns {df_columns}. This will likely make it hard to map columns to code. "
-                        f"To get rid of this warning, use the prefix of the function name to prefix the column names. "
-                        f"e.g. spend_df() would require columns to start with spend_."
-                    )
-                df_dict = output.to_dict(orient="series")
+
+                df_dict = {
+                    get_output_name(name, col_name): col_value
+                    for col_name, col_value in output.to_dict(orient="series").items()
+                }
                 flattened_outputs.update(df_dict)
             elif isinstance(output, pd.Series):
                 if name in flattened_outputs:
