+++ README.md
@@ -1,319 +1,11 @@
-<div align="center">
-    <h1>Welcome to Hamilton's Github Repository</h1>
-    <a href="https://circleci.com/gh/stitchfix/hamilton/tree/main" target="_blank">
-      <img src="https://circleci.com/gh/stitchfix/hamilton/tree/main.svg?style=svg" alt="Hamilton CircleCI"/>
-    </a>
-    <a href="https://join.slack.com/t/hamilton-opensource/shared_invite/zt-1bjs72asx-wcUTgH7q7QX1igiQ5bbdcg" target="_blank"><img src="https://img.shields.io/badge/Join-Hamilton_Slack-brightgreen?logo=slack" alt="Hamilton Slack"/></a>
-        <a href="https://twitter.com/hamilton_os" target="_blank"><img src="https://img.shields.io/twitter/url/http/shields.io.svg?style=social" alt="Twitter"/></a>
-        <br/>
-    <a href="https://www.python.org/downloads/" target="_blank"><img src="https://img.shields.io/badge/python-3.7|%203.8|%203.9|%203.10|%203.11-brightgreen.svg" alt="Python supported"/></a>
-    <a href="https://pypi.org/project/sf-hamilton/" target="_blank"><img src="https://badge.fury.io/py/sf-hamilton.svg" alt="PyPi Version"/></a>
-    <a href="https://pepy.tech/project/sf-hamilton" target="_blank"><img src="https://pepy.tech/badge/sf-hamilton" alt="Total Downloads"/></a>
-    <br/>
-    <a href="https://anaconda.org/hamilton-opensource/sf-hamilton"> <img src="https://anaconda.org/hamilton-opensource/sf-hamilton/badges/version.svg" /> </a>
-    <a href="https://anaconda.org/hamilton-opensource/sf-hamilton"> <img src="https://anaconda.org/hamilton-opensource/sf-hamilton/badges/platforms.svg" /> </a>
-</div>
+# WE HAVE MOVED REPOSITORIES
 
-# Hamilton
+This repository is no longer maintained. Hamilton is now located at https://github.com/DAGWorks-Inc/hamilton. All other aspects of Hamilton (pypi/conda packages, slack community, etc...), have not changed at all.
 
-The general purpose micro-framework for creating dataflows from python functions!
+If you find yourself here, please:
 
-Specifically, Hamilton defines a novel paradigm, that allows you to specify a flow of (delayed) execution, that forms a Directed Acyclic Graph (DAG).
-It was originally built to solve creating wide (1000+) column dataframes. Core to the design of Hamilton is a clear mapping of
-function name to dataflow output. That is, Hamilton forces a certain paradigm with writing functions, and aims for DAG clarity,
-easy modifications, with always unit testable and naturally documentable code.
+1. Give the new repository a star
+2. (if from a link) Open up an issue in the new repository so we can modify the link.
 
-For the backstory on how Hamilton came about, see our [blog post!](https://multithreaded.stitchfix.com/blog/2021/10/14/functions-dags-hamilton/).
-
-# Getting Started
-Here's a quick getting started guide to get you up and running in less than 15 minutes.
-If you need help join our [slack](https://join.slack.com/t/hamilton-opensource/shared_invite/zt-1bjs72asx-wcUTgH7q7QX1igiQ5bbdcg) community to chat/ask Qs/etc.
-For the latest updates, follow us on [twitter](https://twitter.com/hamilton_os)!
-
-## Installation
-Requirements:
-
-* Python 3.7+
-
-To get started, first you need to install hamilton. It is published to pypi under `sf-hamilton`:
-> pip install sf-hamilton
-
-Note: to use the DAG visualization functionality, you should instead do:
-> pip install "sf-hamilton[visualization]"
-
[…]
