+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,17 +1,5 @@
-[Short description explaining the high-level reason for the pull request]
+# THE HAMILTON REPOSITORY HAS MIGRATED
 
-## Changes
+This repository is no longer maintained or monitored.
 
-## How I tested this
-
-## Notes
-
-## Checklist
-
-- [ ] PR has an informative and human-readable title (this will be pulled into the release notes)
-- [ ] Changes are limited to a single goal (no scope creep)
-- [ ] Code passed the pre-commit check & code is left cleaner/nicer than when first encountered.
-- [ ] Any _change_ in functionality is tested
-- [ ] New functions are documented (with a description, list of inputs, and expected output)
-- [ ] Placeholder code is flagged / future TODOs are captured in comments
-- [ ] Project documentation has been updated if adding/changing functionality.
+Please make any pull requests on https://github.com/DAGWorks-Inc/hamilton/.  
