+++ README.md
@@ -5,7 +5,7 @@ This repository is no longer maintained. Hamilton is now located at https://gith
 If you find yourself here, please:
 
 1. [Give the new repository a ⭐](https://github.com/DAGWorks-Inc/hamilton).
-2. (if from a link) Open up an issue in the new repository so we can modify the link.
+2. (if from a link) [Open up an issue in the new repository](https://github.com/DAGWorks-Inc/hamilton/issues) so we can modify the link.
 
 Thank you!
 - Your dedicated Hamilton team
