+++ decorators.md
@@ -1,673 +1,1 @@
-# Decorators
 
-While the 1:1 mapping of column -> function implementation is powerful, we've implemented a few decorators to promote
-business-logic reuse. The decorators we've defined are as follows
-(source can be found in [function_modifiers](hamilton/function_modifiers.py)):
-
-## @parameterize
-Expands a single function into n, each of which correspond to a function in which the parameter value is replaced either by:
-1. A specified value
-2. The value from a specified upstream node.
-
-Note that this can take the place of any of the `@parameterize` decorators below. In fact, they delegate to this!
-
-```python
-import pandas as pd
-from hamilton.function_modifiers import parameterize
-from hamilton.function_modifiers import value, source
-
-
-@parameterize(
-    D_ELECTION_2016_shifted=dict(n_off_date=source('D_ELECTION_2016'), shift_by=value(3)),
-    SOME_OUTPUT_NAME=dict(n_off_date=source('SOME_INPUT_NAME'), shift_by=value(1)),
-)
-def date_shifter(n_off_date: pd.Series, shift_by: int = 1) -> pd.Series:
-    """{one_off_date} shifted by shift_by to create {output_name}"""
-    return n_off_date.shift(shift_by)
-```
-
-By choosing `literal` or `upstream`, you can determine the source of your dependency. Note that you can
-also pass documentation. If you don't, it will use the parameterized docstring.
-
-```python
-@parameterize(
-    D_ELECTION_2016_shifted=(dict(n_off_date=source('D_ELECTION_2016'), shift_by=value(3)), "D_ELECTION_2016 shifted by 3"),
-    SOME_OUTPUT_NAME=(dict(n_off_date=source('SOME_INPUT_NAME'), shift_by=value(1)),"SOME_INPUT_NAME shifted by 1")
-)
-def date_shifter(n_off_date: pd.Series, shift_by: int=1) -> pd.Series:
-    """{one_off_date} shifted by shift_by to create {output_name}"""
-    return n_off_date.shift(shift_by)
-```
-
-
-
-## @parameterize_values (replacing @parametrized)
-Expands a single function into n, each of which corresponds to a function in which the parameter value is replaced by
-that *specific value*.
-```python
-import pandas as pd
-from hamilton.function_modifiers import parameterize_values
-import internal_package_with_logic
-
-ONE_OFF_DATES = {
-     #output name        # doc string               # input value to function
-    ('D_ELECTION_2016', 'US Election 2016 Dummy'): '2016-11-12',
-    ('SOME_OUTPUT_NAME', 'Doc string for this thing'): 'value to pass to function',
-}
-            # parameter matches the name of the argument in the function below
-@parameterize_values(parameter='one_off_date', assigned_output=ONE_OFF_DATES)
-def create_one_off_dates(date_index: pd.Series, one_off_date: str) -> pd.Series:
-    """Given a date index, produces a series where a 1 is placed at the date index that would contain that event."""
-    one_off_dates = internal_package_with_logic.get_business_week(one_off_date)
-    return internal_package_with_logic.bool_to_int(date_index.isin([one_off_dates]))
-```
-We see here that `parameterized` allows you keep your code DRY by reusing the same function to create multiple
-distinct outputs. The _parameter_ key word argument has to match one of the arguments in the function. The rest of
-the arguments are pulled from outside the DAG. The _assigned_output_ key word argument takes in a dictionary of
-tuple(Output Name, Documentation string) -> value.
-
-Note that `@parametrized` is deprecated, and we intend for you to use `@parameterize_vales`. We're consolidating
-to make the parameterization decorators more consistent! You have plenty of time to migrate,
-we wont make this a hard change until we have a Hamilton 2.0.0 to release.
-
-
-## @parameterize_sources (replacing @parameterized_inputs)
-
[…]
