+++ data_quality.md
@@ -1,151 +1,1 @@
-# Data Quality
 
-Hamilton has a simple but powerful data quality capability. This enables you to write functions
-that have assertion on the output. For example...
-
-```python
-import pandas as pd
-import numpy as np
-from hamilton.function_modifiers import check_output
-
-@check_output(
-    data_type=np.int64,
-    data_in_range=(0,100),
-    importance="warn",
-)
-def some_int_data_between_0_and_100() -> pd.Series:
-    pass
-```
-
-In the above, we run two assertions:
-
-1. That the series has an np.int64 datatype
-2. That every item in the series in between 0 and 100
-
-
-Furthermore, the workflow does not fail when this dies. Rather, it logs a warning, as specified by the value provided to `importance`. In terms of how this works, if you were to visualize what was being executed (e.g. using `visualize_exection()`) then you'd see extra nodes added to the DAG. So when using `@check_output` an extra computational step will be added to your workflow to run that check.
-
-
-## Design
-
-To add data quality validation, we run an additional computational step in your workflow after function calculation.
-See comments on the `BaseDataValidationDecorator` class for how it works.
-
-## Default Validators
-
-The available default validators are listed in the variable `AVAILABLE_DEFAULT_VALIDATORS`
-in `default_validators.py`. To add more, please implement the class in that file then add to the list.
-There is a test that ensures that everything is added to that list.
-
-## Pandera Integration
-
-We've fully integrated data quality with [pandera](https://pandera.readthedocs.io/en/stable/)!
-
-Note that you have to have hamilton installed with the `pandera` extension. E.G.
-
-```bash
-pip install sf-hamilton[pandera]
-```
-
-The integration point is simple. All you have to do is provide a pandera schema
-using the default data validator with argument `schema=`. This will validate the
-output against a schema provided by you.
-
-If you don't know what a pandera schema is or haven't worked with them before,
-read more about it [here](https://pandera.readthedocs.io/en/stable/schema_models.html).
-The integration works with schemas for both series and dataframes.
-
-### Validating DataFrames
-
-```python
-import pandera as pa
-import pandas as pd
-from hamilton import function_modifiers
-
-@function_modifiers.check_output(schema=pa.DataFrameSchema(
-        {
-            'column1': pa.Column(int),
-            'column2': pa.Column(float, pa.Check(lambda s: s < -1.2)),
-            # you can provide a list of validators
-            'column3': pa.Column(str, [
-                pa.Check(lambda s: s.str.startswith('value')),
-                pa.Check(lambda s: s.str.split('_', expand=True).shape[1] == 2)
-            ]),
-        },
-        index=pa.Index(int),
-        strict=True,
-    ))
-def dataframe_with_schema(...) -> pd.DataFrame:
-    ...
-```
-
-### Validating Series
-
-```python
-import pandera as pa
-import pandas as pd
-from hamilton import function_modifiers
-
-@function_modifiers.check_output(schema = pa.SeriesSchema(
-        str,
-        checks=[
-            pa.Check(lambda s: s.str.startswith('foo')),
-            pa.Check(lambda s: s.str.endswith('bar')),
-            pa.Check(lambda x: len(x) > 3, element_wise=True)
-        ],
-        nullable=False,
-    ))
-def series_with_schema(...) -> pd.Series:
-    ...
-```
-
-
-You can also do schema checks on series, using the `pa.SeriesSchema` feature!
-
-## Custom Validators
-
-To add a custom validator, you need to implement the class `DataValidator`. You can then use the
-`@check_output_custom` decorator to run it on a function. For example:
-
-```python
-import pandas as pd
-import numpy as np
-
[…]
